Удаляет компонент кнопки play для оптимизации
